# server/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Spraywall API
  version: "0.1.0"
  description: API for the Spraywall climbing app
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /gyms/{gymSlug}/walls/{wallId}/routes:
    get:
      operationId: listRoutes
      summary: List routes on a wall
      tags: [Route]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gymSlug"
        - $ref: "#/components/parameters/wallId"
      responses:
        "200":
          description: List of routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Route"
                required: [routes]
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createRoute
      summary: Create a route
      tags: [Route]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gymSlug"
        - $ref: "#/components/parameters/wallId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequest"
      responses:
        "201":
          description: Route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /gyms/{gymSlug}/walls/{wallId}/routes/{routeId}:
    get:
      operationId: getRoute
      summary: Get route by ID
      tags: [Route]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gymSlug"
        - $ref: "#/components/parameters/wallId"
        - $ref: "#/components/parameters/routeId"
      responses:
        "200":
          description: Route detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /gyms/{gymSlug}/invites:
    post:
      operationId: createInvite
      summary: Create a gym invite
      tags: [Invite]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gymSlug"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /invites/{code}/accept:
    post:
      operationId: acceptInvite
      summary: Accept a gym invite
      tags: [Invite]
      security:
        - BearerAuth: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                required: [status]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    gymSlug:
      name: gymSlug
      in: path
      required: true
      schema:
        type: string
    wallId:
      name: wallId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    routeId:
      name: routeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Route:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wall_id:
          type: string
          format: uuid
        name:
          type: string
        grade:
          type: string
        holds:
          type: array
          items:
            $ref: "#/components/schemas/Hold"
        setter_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      required: [id, wall_id, name]
    Hold:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
        t:
          type: string
          description: "Hold type (start, middle, finish)"
      required: [x, y, t]
    CreateRouteRequest:
      type: object
      properties:
        name:
          type: string
        grade:
          type: string
        holds:
          type: array
          items:
            $ref: "#/components/schemas/Hold"
      required: [name, holds]
    Invite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gym_id:
          type: string
          format: uuid
        code:
          type: string
        expires_at:
          type: string
          format: date-time
      required: [id, gym_id, code]